Add ability to change host, port, path to www directory using command-line arguments, improve the path normalization
This allows the web server to be bound to different hosts/ip's using command-line arguments. This also enables the path of the www directory to be specified on startup. I also fixed the path normalization to additionally strip the leading \ off of the normalized path, this was causing a bug when joining the path of the www directory with the requested path.